requirements.txt
ansitable==0.9.10
cfgv==3.4.0
colorama==0.4.6
colored==1.4.4
contourpy==1.1.1
cycler==0.12.1
distlib==0.3.8
filelock==3.13.3
fonttools==4.50.0
identify==2.5.35
importlib_resources==6.4.0
kiwisolver==1.4.5
loguru==0.7.2
matplotlib==3.7.5
nodeenv==1.8.0
numpy==1.24.4
packaging==24.0
pgraph-python==0.6.2
pillow==10.2.0
platformdirs==4.2.0
pre-commit==3.5.0
progress==1.6
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
retrying==1.3.4
roboticstoolbox-python==1.1.0
rtb-data==1.0.1
scipy==1.10.1
six==1.16.0
spatialgeometry==1.1.0
spatialmath-python==1.1.9
swift-sim==1.1.0
typing_extensions==4.10.0
virtualenv==20.25.1
websockets==12.0
win32-setctime==1.1.0
zipp==3.18.1
